requirements.txt
fastapi
uvicorn[standard]
pydantic
pytz
pandas
openpyxl
google-generativeai
aiosqlite